multiple changes
instead of having 3 lines per checking of dds or png. I wrapped them in a function called DDSorPNG. it's very similar to sprintf.

it'll check for the DDS file if not found it returns the PNG value.

I found one error message treated like it was a file looked like it was copied from battle stages. fixed it.

LoadImageIntoStruct is a function that takes two copies of identical code and makes one function.

changed bActive into a bool

RenderTexture instead of having the same code in like 5 places to check if it's compressed or not.

LoadIamgeFromFile is now const char * const. it's a pointer that isn't changing and the thing it's pointing to doesn't change..