mostly battle scene animation
Added World to Loading image function name to make it different from battle version of function.

Added a new value to struct called nextFrame. This just tracks what frame we are loading next.
current frame being nextFrame - 1;

Changed TRUE to true and Changed BOOL to bool

LoadImageIntoBattleSceneryStruct allows us to call the same code without having to copy and paste it everywhere for battle scenes.

we now check for _0++ till it fails to load and goes back to 0. This method does allow semi unlimited number number of frames.
animation this way hates PNG. Loves DDS.

Fix debut message as was sprintf and then unused.

DDSorPNG now returns false if it found an image. The idea is you stick this in an if statement allowing you to chain calls with out having to do the check outside of the function. So fail return true so you can try something else.