Fix small issues i can find (#36)
* scale variable isn't used.

* update message to print out the image dims as well as the fcp dims

* ignore some cmake crap

* better error when file not found.

* add .gitignore to solution so you can edit in vs

* wip for palette fix.

* remove the /2

* casting to void to hide warnings and add __func__ to debug

* use outputdebug instead of outputdebugstringa

* Update texturepatch_v2_world.cpp

doesn't need path because it's printed when file is opened

* fixed worldmap.

it was using texIndex instead of wmStructPointer

* vs doesn't save automaticlly all my debug lines were still in.

* marked wrong value as static

* RenderTexture

trying to reduce boilerplate by defining RenderTexture a function that calls compressed or uncompressed.

* multiple changes

instead of having 3 lines per checking of dds or png. I wrapped them in a function called DDSorPNG. it's very similar to sprintf.

it'll check for the DDS file if not found it returns the PNG value.

I found one error message treated like it was a file looked like it was copied from battle stages. fixed it.

LoadImageIntoStruct is a function that takes two copies of identical code and makes one function.

changed bActive into a bool

RenderTexture instead of having the same code in like 5 places to check if it's compressed or not.

LoadIamgeFromFile is now const char * const. it's a pointer that isn't changing and the thing it's pointing to doesn't change..

* mostly battle scene animation

Added World to Loading image function name to make it different from battle version of function.

Added a new value to struct called nextFrame. This just tracks what frame we are loading next.
current frame being nextFrame - 1;

Changed TRUE to true and Changed BOOL to bool

LoadImageIntoBattleSceneryStruct allows us to call the same code without having to copy and paste it everywhere for battle scenes.

we now check for _0++ till it fails to load and goes back to 0. This method does allow semi unlimited number number of frames.
animation this way hates PNG. Loves DDS.

Fix debut message as was sprintf and then unused.

DDSorPNG now returns false if it found an image. The idea is you stick this in an if statement allowing you to chain calls with out having to do the check outside of the function. So fail return true so you can try something else.

* update comment

* Update ff8_demaster.vcxproj

return path to what it was.

* redisable water

* fix names of varibles and functions that had changed in refactor