fix function for converting to strings
you were casting char** to char* and returning it.

This is more correct. All c style arrays degrade into pointers. so you didn't need char* infront of the array just char. Moved the global value into the function as a static.

Though the second const might not be needed. it's kinda something you can remove later if it's a problem. it basically means the pointer can't change.