Update texturepatch_v2_world.cpp
I added some code to see about the textureType == 0
so as it had extra texture pages. So I wanted to check if they were different or not.
most of them are probably a variant of the minimap. I decided to leave code there encase we wanted to look closer at it later.

I disabled the noticeable clouds and minimap colliding. Hopefully when you get time to debug this you can figure out a better way to handle them.
right now you can't change one with out the other changing.