Fix crash error (#41)
* fix crash

fixes: https://github.com/MaKiPL/FF8_demastered/issues/40

Thee issue is we were generating a `vector<string>` for all the map names. And we are using the games internal list. But the game does not load the entire list. It only loads as much as needed. So since we have the list cashed as a vector we were going out of range every time we found one that was not in the list.

* adjust code

* remove comments and debug code.

* change to iterate t with a ++

* change t to tmp_ptr and used comma operator.

* added comments and clean up format

* combined the output debug line into one line