fixes
Okay I found and fixed some issues. Seems to work as expected now.
One issue was it was incrementing the frame counter and never setting it back to 1. So i would load a lot more copies of images than needed.
The crash bug was from failing to load a texture because there were too many copies.

So instead crashing you may see a black texture in the case of a failure to load. But it should be less likely to happen now.

demaster.zip
2.83 MB
[10:08 PM]
Though this is something to be concerned with if you go overboard on your animations by adding too many frames you could hit the limit of the computer. And start seeing black textures. Though if we used more dds over png we'd have more headroom. The png is surprisingly usable with animations now because of the fixes to the cache.